refactor: toEntity 위치 변경
